refactor(kitchenpos): create common context and move common domain models to that context
