+++ src/main/java/kitchenpos/common/tobe/DisplayedName.java
@@ -1,4 +1,4 @@
-package kitchenpos.products.tobe.domain;
+package kitchenpos.common.tobe;
 
 import java.util.Objects;
